+++ pepper/rc/changelog.md
@@ -16,6 +16,8 @@
 - added logging to the editor which you can open it with the `open-log` command
 - added `if` command that supports `==` and `!=` operations to conditionally execute other commands
 - added `@platform()` that expands to the platform name the editor is running on top of.
+- added the possibility for plugins to register its own expansions
+- added command variable expansion completion
 
 # 0.27.0
 - added `set-env` command to change the editor's environment variables